% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_create.R
\name{annotation_create}
\alias{annotation_create}
\title{annotation_create}
\usage{
annotation_create(.data, .name_as_pname = FALSE)
}
\arguments{
\item{.data}{A dataframe.}

\item{.name_as_pname}{If TRUE, the column names are also added as the pname.}
}
\value{
A matrix with all names of the input dataset in row 'name' and
additional rows 'pname', 'unit', 'short_pname', and 'comment'.
}
\description{
Creates an annotation file for the input dataset.
}
\examples{
-
}
