% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{check_data}
\usage{
check_data(d, caption = "", print = TRUE)
}
\arguments{
\item{d}{A data.frame or tibble}

\item{caption}{A character vector used as the caption of the output}

\item{print}{If print = TRUE, an HTML output is generated, else a matrix is
returned}

\item{percent_missing}{A number determining the percentage of
missingness used to count variables with high amount of missing values}
}
\value{
Either a HTML output for use in Markdown files or a matrix
}
\description{
Provides a table with info on missingness and variable type for
the input dataset
}
\examples{
-
}
