% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{comment}
\alias{comment}
\title{comment}
\usage{
comment(.comment = "")
}
\arguments{
\item{.comment}{A character vector}
}
\value{
A matrix containing your comments is created in the global
environment
}
\description{
Let's you comment your analysis with one simple function
}
\examples{
-
}
