% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{data_check}
\alias{data_check}
\title{data_check}
\usage{
data_check(.data, .caption = "", .print = TRUE)
}
\arguments{
\item{.data}{A data.frame or tibble}

\item{.caption}{A character vector used as the caption of the output}

\item{.print}{If .print = TRUE, an HTML output is generated, else a matrix is
returned}
}
\value{
Either a HTML output for use in Markdown files or a matrix
}
\description{
Provides a table with info on missingness and variable type for
the input dataset
}
\examples{
-
}
