% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{data_clean}
\usage{
data_clean(
  .data,
  .rows = TRUE,
  .columns = TRUE,
  .row_percent = 100,
  .column_percent = 100,
  .print = TRUE,
  .return_tbl_df = FALSE
)
}
\arguments{
\item{.data}{A data.frame or tibble.}

\item{.rows}{If TRUE, empty rows are omitted.}

\item{.columns}{If TRUE, empty columns are omitted.}

\item{.row_percent}{Integer, indicates the percentage of missing values that
define an empty row.}

\item{.column_percent}{Integer, indicates the percentage of missing values
that define an empty column.}

\item{.print}{If .print = TRUE, a short summary is returned.}

\item{.return_tbl_df}{If TRUE a tibble is returned, else a data.frame is
returned.}
}
\value{
A data.frame or tibble without empty rows and columns.
}
\description{
Returns the input dataframe without empty rows and columns.
}
\examples{
-
}
