% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ready.r
\name{data_ready}
\alias{data_ready}
\title{data_ready}
\usage{
data_ready(.data)
}
\arguments{
\item{.data}{A data.frame or tibble.}
}
\value{
A message indicating how ready your data is for analysis.
}
\description{
Returns a suggestion on how ready your data is for analysis.
}
\examples{
-
}
