% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummies_annotate.R
\name{dummies_annotate}
\alias{dummies_annotate}
\title{dummies_annotate}
\usage{
dummies_annotate(.data, .input_variables)
}
\arguments{
\item{.data}{A data.frame or tibble}

\item{.input_variables}{A character vector containing variables that should be
recoded as dummy variables in the annotation}
}
\value{
The annotation with additional dummy variables for all
.input_variables
}
\description{
Returns the annotation with additional dummy variables for
specified variables. Requires a data.frame or matrix object of name
'annotation' in the global environment. The annotation object must be
specified using the column names "name", "pname", "unit", "short_pname", and
"comment". This function grabs the "pname" of the .input_variables to create
the dummy names.
}
\examples{
-
}
