% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummies_make.R
\name{dummies_make}
\alias{dummies_make}
\title{dummies_make}
\usage{
dummies_make(.data, .input_variables)
}
\arguments{
\item{.data}{A data.frame or tibble}

\item{.input_variables}{A character vector containing variables that should be
recoded as dummy variables}
}
\value{
The a data.frame of the input data with additional dummy variables
for all .input_variables
}
\description{
Returns the input data with additional dummy variables for
specified variables
}
\examples{
-
}
