% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_binary.R
\name{find_binary}
\alias{find_binary}
\title{find_binary}
\usage{
find_binary(
  .data,
  .output = "bin",
  .include = FALSE,
  .yes = NULL,
  .no = NULL,
  .na = NULL,
  .print = TRUE
)
}
\arguments{
\item{.data}{A data.frame or tibble.}

\item{.output}{"bin" returns a vector of binary variables, "list" returns a
list with binary variables, variables coded yes/no and more information.}

\item{.include}{If .include = TRUE, variables recoded as 0/1 within the function
are returned in the list-output - requires .output = "list".}

\item{.yes}{Optional character vector that can be used to feed values that
should be used as 'yes'.}

\item{.no}{Optional character vector that can be used to feed values that
should be used as 'no'.}

\item{.na}{Optional character vector that can be used to feed values that
should be used as 'NA'.}

\item{.print}{If .print = TRUE, values detected as yes/no are
printed.}
}
\value{
Either a character vector with clearly binary variables or a list
with additional information.
}
\description{
Returns either a vector with all clearly binary variables or a
list with additional information on variables that might also be binary.
}
\examples{
-
}
