% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummies.R
\name{make_dummies}
\alias{make_dummies}
\title{make_dummies}
\usage{
make_dummies(d, input_variables)
}
\arguments{
\item{d}{A data.frame or tibble}

\item{input_variables}{A character vector containing variables that should be
recoded as dummy variables}
}
\value{
The a data.frame of the input data with additional dummy variables
for all input_variables
}
\description{
Returns the input data with additional dummy variables for
specified variables
}
\examples{
-
}
