% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_structure.R
\name{project_structure}
\alias{project_structure}
\title{project_structure}
\usage{
project_structure(.path = "")
}
\arguments{
\item{.path}{A character vector containing the path to the project.}
}
\value{
A character vector with the numeric variables.
}
\description{
Creates a project structure in the current working directory.
}
\examples{
-
}
