% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_na.R
\name{show_na}
\alias{show_na}
\title{show_na}
\usage{
show_na(d, caption = "", print = TRUE)
}
\arguments{
\item{d}{A data.frame or tibble}

\item{caption}{A character vector used as the caption of the output}

\item{print}{If print = TRUE, an HTML output is generated, else a matrix is
returned}
}
\value{
Either a HTML output for use in Markdown files or a matrix
}
\description{
Provides an overview on missing values for the input dataset
}
\examples{
-
}
