% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_start.R
\name{time_start}
\alias{time_start}
\title{time_start}
\usage{
time_start()
}
\value{
Creates a time_started object in the global environment.
}
\description{
Initializes the starting point of a time measurement.
}
\examples{
-
}
