% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_stop.R
\name{time_stop}
\alias{time_stop}
\title{time_stop}
\usage{
time_stop()
}
\value{
Creates a 'time_it_took' object in the global environment.
}
\description{
Calculates the time difference since 'time_start' has been
initialized.
}
\examples{
-
}
