% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_structure.R
\name{project_structure}
\alias{project_structure}
\title{project_structure}
\usage{
project_structure(.path = "", .template = "basic", .data_folder = TRUE)
}
\arguments{
\item{.path}{A character vector containing the path to the project.}

\item{.template}{Defaults to "basic". Character vector that can be used to
choose user-defined templates that have been saved to
"system.file/templates".}

\item{.data_folder}{Logical. If TRUE, a folder of name 'data' with
subdirectories 'raw' and 'clean' is created in the root folder of the current
project.}
}
\value{
A character vector with the numeric variables.
}
\description{
Creates a project structure in the current working directory.
}
\examples{
-
}
