% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_na.R
\name{show_na}
\alias{show_na}
\title{show_na}
\usage{
show_na(.data, .caption = "", .print = TRUE)
}
\arguments{
\item{.data}{A data.frame or tibble}

\item{.caption}{A character vector used as the caption of the output}

\item{.print}{If .print = TRUE, an HTML output is generated, else a matrix is
returned}
}
\value{
Either a HTML output for use in Markdown files or a matrix
}
\description{
Returns overview on missing values for variables of the input
dataset containing at least one missing value
}
\examples{
-
}
